+++ torch_geometric/nn/pool/select/__init__.py
@@ -1,3 +1,8 @@
 from .base import Select, SelectOutput
+from .topk_select import TopkSelect
 
-__all__ = ['Select', 'SelectOutput']
+__all__ = [
+    'Select',
+    'SelectOutput',
+    'TopkSelect',
+]
