+++ torch_geometric/nn/pool/select/topk_select.py
@@ -1,0 +1,87 @@
+from typing import Optional, Union
+
+import torch
+from torch import Tensor
+
+from torch_geometric.utils import scatter
+
+from .base import Select, SelectOutput
+
+
+def topk(
+    x: Tensor,
+    ratio: Optional[Union[float, int]],
+    batch: Tensor,
+    min_score: Optional[float] = None,
+    tol: float = 1e-7,
+) -> Tensor:
+    if min_score is not None:
+        # Make sure that we do not drop all nodes in a graph.
+        scores_max = scatter(x, batch, reduce='max')[batch] - tol
+        scores_min = scores_max.clamp(max=min_score)
+
+        perm = (x > scores_min).nonzero().view(-1)
+
+    elif ratio is not None:
+        num_nodes = scatter(batch.new_ones(x.size(0)), batch, reduce='sum')
+        batch_size, max_num_nodes = num_nodes.size(0), int(num_nodes.max())
+
+        cum_num_nodes = torch.cat(
+            [num_nodes.new_zeros(1),
+             num_nodes.cumsum(dim=0)[:-1]], dim=0)
+
+        index = torch.arange(batch.size(0), dtype=torch.long, device=x.device)
+        index = (index - cum_num_nodes[batch]) + (batch * max_num_nodes)
+
+        dense_x = x.new_full((batch_size * max_num_nodes, ), -60000.0)
+        dense_x[index] = x
+        dense_x = dense_x.view(batch_size, max_num_nodes)
+
+        _, perm = dense_x.sort(dim=-1, descending=True)
+
+        perm = perm + cum_num_nodes.view(-1, 1)
+        perm = perm.view(-1)
+
+        if ratio >= 1:
+            k = num_nodes.new_full((num_nodes.size(0), ), int(ratio))
+            k = torch.min(k, num_nodes)
+        else:
+            k = (float(ratio) * num_nodes.to(x.dtype)).ceil().to(torch.long)
+
+        if isinstance(ratio, int) and (k == ratio).all():
+            # If all graphs have exactly `ratio` or more than `ratio` entries,
+            # we can just pick the first entries in `perm` batch-wise:
+            index = torch.arange(batch_size, device=x.device) * max_num_nodes
+            index = index.view(-1, 1).repeat(1, ratio).view(-1)
+            index += torch.arange(ratio, device=x.device).repeat(batch_size)
+        else:
+            # Otherwise, compute indices per graph:
+            index = torch.cat([
+                torch.arange(k[i], device=x.device) + i * max_num_nodes
+                for i in range(batch_size)
+            ], dim=0)
+
+        perm = perm[index]
+
+    else:
+        raise ValueError("At least one of 'min_score' and 'ratio' parameters "
+                         "must be specified")
+
+    return perm
+
+
+class TopkSelect(Select):
+    def __init__(self, ratio: Optional[float],
+                 min_score: Optional[float] = None, tol: float = 1e-7):
+        super().__init__()
+        self.ratio = ratio
+        self.min_score = min_score
+
+    def forward(self, x: Tensor, batch: Tensor) -> SelectOutput:
+        perm = topk(x, self.ratio, batch, self.min_score)
+        num_clusters = perm.size(0)
+        weight = torch.ones_like(perm, dtype=torch.float)
+        return SelectOutput(node_index=perm,
+                            cluster_index=torch.arange(num_clusters),
+                            num_clusters=num_clusters, num_nodes=x.size(0),
+                            weight=weight)
