+++ torch_geometric/nn/pool/topk_pool.py
@@ -6,26 +6,9 @@ from torch.nn import Parameter
 
 from torch_geometric.nn.inits import uniform
 from torch_geometric.nn.pool.connect import Connect
-from torch_geometric.nn.pool.select import Select, SelectOutput
-from torch_geometric.utils import scatter, softmax
+from torch_geometric.nn.pool.select import SelectOutput, TopkSelect
+from torch_geometric.utils import softmax
 from torch_geometric.utils.num_nodes import maybe_num_nodes
-
-
-class TopkSelect(Select):
-    def __init__(self, ratio: Optional[float],
-                 min_score: Optional[float] = None, tol: float = 1e-7):
-        super().__init__()
-        self.ratio = ratio
-        self.min_score = min_score
-
-    def forward(self, x: Tensor, batch: Tensor) -> SelectOutput:
-        perm = topk(x, self.ratio, batch, self.min_score)
-        num_clusters = perm.size(0)
-        weight = torch.ones_like(perm, dtype=torch.float)
-        return SelectOutput(node_index=perm,
-                            cluster_index=torch.arange(num_clusters),
-                            num_clusters=num_clusters, num_nodes=x.size(0),
-                            weight=weight)
 
 
 class TopkReduce(torch.nn.Module):
@@ -47,69 +30,8 @@ class TopkConnect(Connect):
             edge_index,
             edge_attr,
             cluster.node_index,
+            cluster.num_nodes,
         )
-
-
-def topk(
-    x: Tensor,
-    ratio: Optional[Union[float, int]],
-    batch: Tensor,
-    min_score: Optional[float] = None,
-    tol: float = 1e-7,
-) -> Tensor:
-    if min_score is not None:
-        # Make sure that we do not drop all nodes in a graph.
-        scores_max = scatter(x, batch, reduce='max')[batch] - tol
-        scores_min = scores_max.clamp(max=min_score)
-
-        perm = (x > scores_min).nonzero().view(-1)
-
-    elif ratio is not None:
-        num_nodes = scatter(batch.new_ones(x.size(0)), batch, reduce='sum')
-        batch_size, max_num_nodes = num_nodes.size(0), int(num_nodes.max())
-
-        cum_num_nodes = torch.cat(
-            [num_nodes.new_zeros(1),
-             num_nodes.cumsum(dim=0)[:-1]], dim=0)
-
-        index = torch.arange(batch.size(0), dtype=torch.long, device=x.device)
-        index = (index - cum_num_nodes[batch]) + (batch * max_num_nodes)
-
-        dense_x = x.new_full((batch_size * max_num_nodes, ), -60000.0)
-        dense_x[index] = x
-        dense_x = dense_x.view(batch_size, max_num_nodes)
-
-        _, perm = dense_x.sort(dim=-1, descending=True)
-
-        perm = perm + cum_num_nodes.view(-1, 1)
-        perm = perm.view(-1)
-
-        if ratio >= 1:
-            k = num_nodes.new_full((num_nodes.size(0), ), int(ratio))
-            k = torch.min(k, num_nodes)
-        else:
-            k = (float(ratio) * num_nodes.to(x.dtype)).ceil().to(torch.long)
-
-        if isinstance(ratio, int) and (k == ratio).all():
-            # If all graphs have exactly `ratio` or more than `ratio` entries,
-            # we can just pick the first entries in `perm` batch-wise:
-            index = torch.arange(batch_size, device=x.device) * max_num_nodes
-            index = index.view(-1, 1).repeat(1, ratio).view(-1)
-            index += torch.arange(ratio, device=x.device).repeat(batch_size)
-        else:
-            # Otherwise, compute indices per graph:
-            index = torch.cat([
-                torch.arange(k[i], device=x.device) + i * max_num_nodes
-                for i in range(batch_size)
-            ], dim=0)
-
-        perm = perm[index]
-
-    else:
-        raise ValueError("At least one of 'min_score' and 'ratio' parameters "
-                         "must be specified")
-
-    return perm
 
 
 def filter_adj(
@@ -208,6 +130,7 @@ class TopKPooling(torch.nn.Module):
         self.nonlinearity = nonlinearity
         self.select = TopkSelect(ratio, min_score)
         self.reduce = TopkReduce(multiplier)
+        self.connect = TopkConnect()
         self.weight = Parameter(torch.Tensor(1, in_channels))
 
         self.reset_parameters()
@@ -255,8 +178,8 @@ class TopKPooling(torch.nn.Module):
         x = self.reduce(x, perm, score)
         batch = batch[perm]
 
-        edge_index, edge_attr = filter_adj(edge_index, edge_attr, perm,
-                                           num_nodes=score.size(0))
+        edge_index, edge_attr = self.connect(select_output, edge_index,
+                                             edge_attr)
 
         return x, edge_index, edge_attr, batch, perm, score[perm]
